Moved utils to utils/mod.rs. Moved dev-deps into correct place. Made const for magic numbers.
