+++ Cargo.toml
@@ -13,7 +13,6 @@ async-trait = "0.1.57"
 tokio = { version = "1", features = ["full"] }
 
 # Web
-jsonrpc = "0.14.0"
 hyper = "0.14.4"
 hyper-tls = "0.5.0"
 tap = "1.0"
@@ -56,6 +55,7 @@ chrono = { version = "0.4", features = ["serde"] }
 
 
 [dev-dependencies]
+jsonrpc = "0.14.0"
 test-context = "0.1"
 
 [build-dependencies]
