+++ tests/functional/connection/mod.rs
@@ -1,2 +1,4 @@
 pub(crate) mod infura;
 pub(crate) mod pokt;
+
+const INFURA_CHAIN_DECOMISSIONED_ERROR_CODE: i32 = -32601;
